Allow session users to access dashboard by role
